AC-1604 sparkql built-in conjunction evaluator
Here is a rewrite of the processing for the boolean algebra in sparkql evaluation. It takes a poorly modeled implementation that tracks nesting levels, to one that only cares about sparkql 'block_groups'. It turns out, these groups can be evaluated in a very consistent manner, and cleaned up this logic a ton.

My first approach involved rebuilding the AST. That started to look expensive, and the original implementation kept things nice and compact, so I went back that direction, even though it's a little harder to debug.

As far as my testing goes, this solution works for all the things the api supports, including the reported problems, and some dropped field resolution we were not doing before.