Added round(), substring(), trim(), numeric negate
This commit is adding the following functions:

round(decimal):
returns the nearest integer

substring(character, integer[, integer]):
starting 0 based index and ending after the optional third parameter
which is the number of characters
Returns a character as a substring from the original.

trim(character):
Trims the leading and trailing whitespace of a character.
Returns a character.

Bumped version to 1.1.14