Nested Field Function support
Field funcions now create a nested attribute named field_manipulations
within the each expression. That attribute will have a `:function` `:type`
attribute and an `:args` attribute. The previous symbols which were used
will be built by the highest leveled function within that hash. Note
that each arg in the `:args` attribute can be a function. We now expect
those to be handled recursively. The previous attributes are kept for
backward compatibility with other departments to prevent breakage.