API-7828 rubocop (#75)
* API-7828 Timezone accounting for time -> date comparisons

After taking a few cracks at this, I've opted to do a few things:

* reorganize the function resolver so it's time / timezone agnostic
* set a parser current_timestamp, and the ability to query the "parser"
timezone. This will allow us to localize the timezone when needed
* localize all query timestamps to the parser time zone
* ensure dates reflect the appropriate timezone when extracted from a
timestamp

* rubocop all the thingz

This was a pretty big overhaul long overdue. I didn't make any changes
that would break interfaces, but everything else was game to get rid of
warnings and security problem reports.

* nit

* API-7828 remove deprecated DateTime references

At least, as much as is reasonable. On instance remains for the year() / month() functionality for simplicity sake.

* match up gem with API

* match up gem with API

* versioning