API-9331 additional fixes
While integrating this fix with our projects, this regression popped up. I've added the same tests to this project.

The bug was very obvious once I saw it: we need to pop the block group from the current level into the expressions for the next level up. I accidentally copied the bottom expressions all the way up. Shocked tests passed with this, honestly.